adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.14.0",
]

[tool.uv.workspace]
members = ["packages/*"]
exclude = ["packages/seeds"]
